refactor: split compiler library into modular files
